requirements.txt
Flask
openai>=1.0.0
requests
python-dotenv
